Refactored after Code Review
